refactor: extract action utils & layout ui
