+++ src/features/dashboard/page-layout.tsx
@@ -84,7 +84,7 @@ function DesktopContent({
               'relative flex h-fit w-full max-w-[1200px] pb-2',
               classNames?.frameWrapper
             ),
-            frame: className,
+            frame: cn(className, 'bg-bg-100'),
           }}
         >
           {children}
