+++ src/server/auth/auth-actions.ts
@@ -6,7 +6,8 @@ import { headers } from 'next/headers'
 import { redirect } from 'next/navigation'
 import { Provider } from '@supabase/supabase-js'
 import { AUTH_URLS, PROTECTED_URLS } from '@/configs/urls'
-import { actionClient, returnServerError } from '@/lib/clients/action'
+import { actionClient } from '@/lib/clients/action'
+import { returnServerError } from '@/lib/utils/action'
 import { z } from 'zod'
 import { zfd } from 'zod-form-data'
